publish: add 'skipCycloneDxBom' parameter
This change allows the publish mojo to skip the building and attaching
of a CycloneDX BOM artifact to the project if the parameter
`skipCycloneDxBom` is set to true (default value is false).

Related to [#59][0], generating this artifact fails for projects where
the parent pom is not in Maven Central, so adding this property is
mostly a workaround to the failure but might have other uses as well.

[0]: https://github.com/openrewrite/rewrite-maven-plugin/issues/59