Upping maven-filter from 3.2.0 (Aug 2020) to 3.3.0 (Jun, 2022) (#384)
Fixing a null value in the activeRecipes collection causing the discover goal to fail even if no recipes should be run. This can happen when having a property as the value of the recipe name and that property is null. Then the collection will only contain a single element which is null

Co-authored-by: Jonas Fredin <jonas.fredin@fortnox.se>