Improve quark parsing message
Fixes #671.

As suggested in the discussion, change the message to indicate that
quark files are not parsed.

Also change the unit from "Mb" to "MB", since "Mb" indicates Megabit.
Add a blank between numbers and unit for readability and to be
consistent with output in other parts of the code.